Use `SCMRevision` instead of `BuildData` to determine revision to notify
When using a Global library that also comes from Bitbucket the notification will sometimes go to that library rather than the revision that the build is running for.
This is because there are multiple `BuildData` actions when retrieving the library.
This change hopes to fix that for Git libraries by retrieving the `SCMRevision` instead of the `BuildData`.

issue: https://issues.jenkins-ci.org/browse/JENKINS-41996
issue: https://issues.jenkins-ci.org/browse/JENKINS-42878