Use ConnectionManager in BitbucketServerAPIClient
Fixes JENKINS-46197
While BitbucketServerAPIClient releases connections, the javadoc on
releaseConnection states:
'Releases the connection being used by this HTTP method. In particular
the
connection is used to read the response(if there is one) and will be
held
until the response has been read. If the connection can be reused by
other
HTTP methods it is NOT closed at this point.'
Using a connection pool prevents exessive connection usage.