5 second ConnectionKeepAliveStrategy
The default behaviour of `PoolingHttpClientConnectionManager` is to inspect the server-provided `Keep-Alive` header and use it as the basis for auto configuring the `ConnectionKeepAliveStrategy`, and in the absence of such a header assumes the connection can be kept alive indefinitely.

api.bitbucket.org does not produce a `Keep-Alive` header.

Hardcode a more aggressive 5-second ConnectionKeepAliveStrategy.

Fixes #217