Shorten a bit test names to avoid Linux encryption issues
This change is a bit weird, I know.
In some cases, the file-length maximum length on Linux can be 143 chars
depending on how you encrypted your disk.
https://stackoverflow.com/questions/6571435/limit-on-file-name-length-in-bash

The problem is the following: we have a way to designate tests in our
internal testing pipelines.
And `com.cloudbees.jenkins.plugins.bitbucket.SSHCheckoutTraitTest.given__sshCheckoutWithCredentials__when__decoratingHg__then__credentialsApplied.txt`
is actually 144 characters-long.

So removing the double `_` is making this filename go to 139 chars,
which is below the max 143.

I understand this change may look special, but I think it's still reasonable
as we keep the code readability while addressing a problem.

Thanks for considering it! :-)