[JENKINS-61119] Configure ForkPullRequestDiscoveryTrait using Job DSL
JENKINS-26535 describes a workaround for this issue by reducing the
generics of the databound constructor. A number of other plugins
blocked by this issue have solved the issue using this method as well.

For this plugin a new abstract class named BitbucketForkTrustPolicy
was introduced and used in the databound constructor. All the trust
policies have been updated to extend this abstract class.

Fixes #290, JENKINS-61119