Decode webhook request body as UTF-8
When receiving a webhook request at bitbucket-scmsource-hook and
converting the HTTP request body to a string, always use UTF-8
even if the system default charset is different.
This is mainly relevant to a Jenkins controller running on Windows,
where the system charset might be Windows-1252, but Bitbucket Server
encodes the webhook request JSON in UTF-8.

Fix <https://github.com/jenkinsci/bitbucket-branch-source-plugin/issues/762>