Fix 'missing the Job/Configure permssion' error on Branch Sources page
When this plugin is used in combination with the "Role-Based
Authorization Strategy", the "Branch Sources" job configation page
throws the following error:

    $USER is missing the Job/Configure permission

This error is shown even when the user has full Admin rights on the
Jenkins instance.

I'm not sure that this fix is 100% correct, but it does resolve the
problem. I noticed that the following PR in a similar project fixed a
similar issue (JENKINS-60116) by checking for the `EXTENDED_READ`
permission:

https://github.com/jenkinsci/atlassian-bitbucket-server-integration-plugin/pull/165

So I blindly did the same thing. Would appreciate someone checking my
work!