Use CANCELLED build state in Bitbucket Data Center and Server 8.0 (#606)
Add an options to BitbucketBuildStatusNotificationsTrait to determine which notification to send to Bitbucket based on the Jenkins build result.
Revert the default behaviour for NOT_BUILT notification from SUCCESS to FAILED (configurable to STOPPED) for Bitbucket Cloud, the SUCCESS status was erroneously introduced in 2022 as default instead of STOPPED with the disableNotificationForNotBuildJob (which can be misunderstood) configurable option.
Add help description for all current configurable options for the build status notification trait