[JENKINS-73250] The GIT_URL environmente variable contains user name after update
Revert to the behavior before the clone link feature from version 866.
The remote URL must not contain the user or it may not work when using username/password authentication. This is because the username belonging to the credential provided to perform the git operation may not be the same as the URL. The clone URL comes from the links that bitbucket sends with the repository information. The link sent contains the user associated with the credentials used to call the REST APIs.