[JENKINS-75157] Bitbucket client fail to retrieve resource with HTTP 404 for bitbucket server (#970)
Remove the HttpHost parameter when calling in Apache client as the request could be different than the host configured in jenkins, for example in Bitbucket Server request using mirror link.
Rethrow the FileNotFoundException in executeMethod instead of encapsulate into other exception to respect the SCMFile interface.
Rewrote the BitbucketSCMFile logic to not assume the SCMFile.type but resolve at runtime when needed.