[JENKINS-75523] PRs created from tags cannot be checked out
Add a new branchType field in the pull request bean that is used to identify in case of pull requests for which the source HEAD is a tag instead of branch. This kind of pull request are available only in Bitbucket DataCenter.
Failure during checkout of PR jobs from HEAD tag has been resolved using as source reference "+refs/tags/<tag_name>:..." instead of "+refs/heads/<tag_name>:...".